Renamed post API
