+++ src/api/post/post.api.ts
@@ -1,5 +1,5 @@
-import { wretchGet } from "./ExternalApi";
-import { PostTO } from "./api.types";
+import { wretchGet } from "../ExternalApi";
+import { PostTO } from "../api.types";
 
 type GetPostParams = {
   id: number;
