+++ src/mocks/browser.ts
